+++ src/main/java/com/softserve/travelagency/service/UserService.java
@@ -10,9 +10,9 @@ public interface UserService {
 
     public void saveUser(User user);
 
-    public User getUserById(int id);
+    public User getUserById(Long id);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
 //    public void deleteUser2(User user);
 }
